==> 1_build_data_recession_pred.ipynb <==
## Targets is S&P500 returns, input is either from FRED St. Louis or Yahoo finance

import numpy as np
import pandas as pd

#### Read in the targets. Periods of downturn are periods where consecutive monthly returns <-15%

df = pd.read_csv('targets.csv')
df = df[['Date','Open','Close','monthly_return','period of downturn','Recessions']]
df['Date'] = pd.to_datetime(df['Date'])
df.head()

#### Read in Industrial Production from FRED STL, lagging by 2 months for results to come out

indpro = pd.read_csv('INDPRO.csv')
indpro.columns = ['Date','INDPRO']
indpro['Date'] = pd.to_datetime(indpro['Date'])

df = df.merge(indpro, how='left', on='Date')
df['INDPRO'] = df['INDPRO'].shift(2)

#### Read in daily yield curve and match with the monthly stock result using fuzzy logic (yc data was available for weekdays only)

def nearest(items, pivot):
    return min(items, key=lambda x: abs(x - pivot))

yield_curve = pd.read_csv('T10Y2Y.csv')
yield_curve.columns = ['Date','YIELD_CURVE']
yield_curve['Date'] = pd.to_datetime(yield_curve['Date'])
yield_curve['YIELD_CURVE'] = np.where(yield_curve['YIELD_CURVE']=='.',None,yield_curve['YIELD_CURVE'])
yield_curve['YIELD_CURVE'].fillna(method='ffill', inplace=True)

df['YIELD_CURVE'] = df['Date'].map(lambda x:nearest(yield_curve['Date'],x))    



df.rename(columns={'YIELD_CURVE':'yc_date'},inplace=True)

#### Accounting for the yc data starting in 1976

df1 = df.merge(yield_curve,how='left',left_on='yc_date',right_on='Date')
df1['YIELD_CURVE'] = np.where(df1['Date_y']==pd.to_datetime('1976-06-01'),None,df1['YIELD_CURVE'])

#### Read in unemployment and lag by 2 months for announcement time. Also checkpoint because fuzzy matching took some time

df1.drop(columns=['Date_y'],inplace=True)
df1.rename(columns={'Date_x':'Date'},inplace=True)

unemploy = pd.read_csv('UNRATE.csv')
unemploy.columns = ['Date','UNEMPLOYMENT']
unemploy['Date'] = pd.to_datetime(unemploy['Date'])

df1 = df1.merge(unemploy, how='left', on='Date')
df1['UNEMPLOYMENT'] = df1['UNEMPLOYMENT'].shift(2)

df1.to_csv('zzz_df_checkpoint.csv',index=False)

#### Read in checkpoint_df

df1 = pd.read_csv('zzz_df_checkpoint.csv')
df1['Date'] = pd.to_datetime(df1['Date'])

#### Read in individual stock data, get difference from 2-month high and merge

tickers = ['BGI','KO','KR','MAR','PPG','SHW']
for tick in tickers:
    new_df = pd.read_csv(tick+'.csv')
    new_df.columns = ['Date','Open','High','Low',tick+'_Close','Adj Close','Volume']
    new_df = new_df[['Date',tick+'_Close']]
    new_df['Date'] = pd.to_datetime(new_df['Date'])
    new_df[tick+'_2_mth_high'] = new_df['Date'].map(lambda x:max(new_df[(new_df['Date']<=pd.to_datetime(x))&(new_df['Date']>=pd.to_datetime(x)-pd.Timedelta(35,'d'))][tick+'_Close']))
    new_df[tick+'_diff_from_2_month_high'] = (new_df[tick+'_Close'] - new_df[tick+'_2_mth_high'])/new_df[tick+'_2_mth_high']
    new_df = new_df[['Date',tick+'_diff_from_2_month_high']]
    
    df1 = df1.merge(new_df,how='left',on='Date')
    df1[tick+'_diff_from_2_month_high'] = df1[tick+'_diff_from_2_month_high'].shift(1)

#### Create difference from 6 month high(low) for unemployment and ind pro

df1 = df1[df1['Date']!=pd.to_datetime('1950-01-01')]

df1['unemployment_6_mth_low'] = df1['Date'].map(lambda x:min(df1[(df1['Date']<=pd.to_datetime(x))&(df1['Date']>=pd.to_datetime(x)-pd.Timedelta(180,'d'))]['UNEMPLOYMENT']))
df1['unemployment_diff_from_6_month_low'] = df1['UNEMPLOYMENT'] - df1['unemployment_6_mth_low']

df1['INDPRO_6_mth_high'] = df1['Date'].map(lambda x:max(df1[(df1['Date']<=pd.to_datetime(x))&(df1['Date']>=pd.to_datetime(x)-pd.Timedelta(180,'d'))]['INDPRO']))
df1['INDPRO_diff_from_6_month_high'] = df1['INDPRO_6_mth_high'] - df1['INDPRO']


df1.drop(columns=['unemployment_6_mth_low','INDPRO_6_mth_high'],inplace=True)
df1.head()

df1.to_csv('reg_ready.csv',index=False)
