==> 2_predicting_downturns_with_rf.ipynb <==
import numpy as np
import pandas as pd

from sklearn.ensemble import RandomForestRegressor,RandomForestClassifier

df = pd.read_csv('reg_ready.csv')
df['Date'] = pd.to_datetime(df['Date'])
df.head()

#### First, rf regression with [all features but MAR and BGI]. Runs back to 1980 or so 

df.columns

REL_VARS = ['YIELD_CURVE','KO_diff_from_2_month_high',
       'KR_diff_from_2_month_high',
       'PPG_diff_from_2_month_high', 'SHW_diff_from_2_month_high',
       'unemployment_diff_from_6_month_low', 'INDPRO_diff_from_6_month_high']

df_train = df[df['Date']<pd.to_datetime('2005-01-01')]
df_test = df[df['Date']>=pd.to_datetime('2005-01-01')]

df_train.dropna(subset=REL_VARS,inplace=True)
df_test.dropna(subset=REL_VARS,inplace=True)

x_train = df_train[REL_VARS]
x_test = df_test[REL_VARS]

y_train = df_train['monthly_return']
y_test = df_test['monthly_return']


rf = RandomForestRegressor(n_estimators=2500,max_depth=4,max_features='sqrt')
rf.fit(x_train,y_train)

print(rf.score(x_train,y_train))

df_test['pred'] = rf.predict(x_test)

print(rf.score(x_test,y_test))

df_test.to_csv('testset_returns_regression_1980.csv',index=False)

#### Classification from 1980

REL_VARS = ['YIELD_CURVE','KO_diff_from_2_month_high',
       'KR_diff_from_2_month_high',
       'PPG_diff_from_2_month_high', 'SHW_diff_from_2_month_high',
       'unemployment_diff_from_6_month_low', 'INDPRO_diff_from_6_month_high']

df_train = df[df['Date']<pd.to_datetime('2005-01-01')]
df_test = df[df['Date']>=pd.to_datetime('2005-01-01')]

df_train.dropna(subset=REL_VARS,inplace=True)
df_test.dropna(subset=REL_VARS,inplace=True)

x_train = df_train[REL_VARS]
x_test = df_test[REL_VARS]

y_train = df_train['period of downturn']
y_test = df_test['period of downturn']


rf = RandomForestClassifier(n_estimators=2500,max_depth=4,max_features='sqrt')
rf.fit(x_train,y_train)

print(rf.score(x_train,y_train))
print(rf.score(x_test,y_test))

df_train['pred'] = rf.predict_proba(x_train)[:,1]

df_test['pred'] = rf.predict_proba(x_test)[:,1]

df_train.to_csv('trainset_returns_classification_1980.csv',index=False)
df_test.to_csv('testset_returns_classification_1980.csv',index=False)

#### Let us try IND_PRO, UNEMPLOYMENT and Coca-Cola (from 1962) reg and classification below

REL_VARS = ['KO_diff_from_2_month_high',
       'unemployment_diff_from_6_month_low', 'INDPRO_diff_from_6_month_high']

df_train = df[df['Date']<pd.to_datetime('2005-01-01')]
df_test = df[df['Date']>=pd.to_datetime('2005-01-01')]

df_train.dropna(subset=REL_VARS,inplace=True)
df_test.dropna(subset=REL_VARS,inplace=True)

x_train = df_train[REL_VARS]
x_test = df_test[REL_VARS]

y_train = df_train['monthly_return']
y_test = df_test['monthly_return']


rf = RandomForestRegressor(n_estimators=2500,max_depth=4,max_features='sqrt')
rf.fit(x_train,y_train)

print(rf.score(x_train,y_train))

df_test['pred'] = rf.predict(x_test)

print(rf.score(x_test,y_test))

df_test.to_csv('testset_returns_regression_1962.csv',index=False)

REL_VARS = ['KO_diff_from_2_month_high',
       'unemployment_diff_from_6_month_low', 'INDPRO_diff_from_6_month_high']

df_train = df[df['Date']<pd.to_datetime('2005-01-01')]
df_test = df[df['Date']>=pd.to_datetime('2005-01-01')]

df_train.dropna(subset=REL_VARS,inplace=True)
df_test.dropna(subset=REL_VARS,inplace=True)

x_train = df_train[REL_VARS]
x_test = df_test[REL_VARS]

y_train = df_train['period of downturn']
y_test = df_test['period of downturn']


rf = RandomForestClassifier(n_estimators=2500,max_depth=4,max_features='sqrt')
rf.fit(x_train,y_train)

print(rf.score(x_train,y_train))
print(rf.score(x_test,y_test))

df_train['pred'] = rf.predict_proba(x_train)[:,1]

df_test['pred'] = rf.predict_proba(x_test)[:,1]

df_train.to_csv('trainset_returns_classification_1962.csv',index=False)
df_test.to_csv('testset_returns_classification_1962.csv',index=False)

#### 1980 was better (more features available in the dataset)

#### Let us visualize feature importance in the 1980 example

REL_VARS = ['YIELD_CURVE',
        'KO_diff_from_2_month_high',
       'KR_diff_from_2_month_high',
       'PPG_diff_from_2_month_high', 
        'SHW_diff_from_2_month_high',
       'unemployment_diff_from_6_month_low', 
        'INDPRO_diff_from_6_month_high']

df_train = df[df['Date']<pd.to_datetime('2005-01-01')]
df_test = df[df['Date']>=pd.to_datetime('2005-01-01')]

df_train.dropna(subset=REL_VARS,inplace=True)
df_test.dropna(subset=REL_VARS,inplace=True)

x_train = df_train[REL_VARS]
x_test = df_test[REL_VARS]

y_train = df_train['period of downturn']
y_test = df_test['period of downturn']


rf = RandomForestClassifier(n_estimators=2500,max_depth=4,max_features='sqrt')
rf.fit(x_train,y_train)

print("let's see the feature importances:")
print(rf.feature_importances_)

#### Let's get rid of the bottom three features and rerun (results saved this time.)

REL_VARS = ['YIELD_CURVE',
       'PPG_diff_from_2_month_high',
       'unemployment_diff_from_6_month_low', 
        'INDPRO_diff_from_6_month_high']

df_train = df[df['Date']<pd.to_datetime('2005-01-01')]
df_test = df[df['Date']>=pd.to_datetime('2005-01-01')]
print(df_test.shape)

df_train.dropna(subset=REL_VARS,inplace=True)
df_test.dropna(subset=REL_VARS,inplace=True)

x_train = df_train[REL_VARS]
x_test = df_test[REL_VARS]

y_train = df_train['period of downturn']
y_test = df_test['period of downturn']


rf = RandomForestClassifier(n_estimators=2500,max_depth=4,max_features='sqrt')
rf.fit(x_train,y_train)

print("let's see the feature importances:")
print(rf.feature_importances_)

df_train['pred'] = rf.predict_proba(x_train)[:,1]

df_test['pred'] = rf.predict_proba(x_test)[:,1]

#df_train.to_csv('trainset_returns_classification_1980.csv',index=False)
#df_test.to_csv('testset_returns_classification_1980.csv',index=False)

#### Ok so yield curve, PPG, unemployment and Industrial Prod have some signal. I think a sequence-based modeling technique will allow us to uncover the point of "first-trouble" that holds a lot of the signal in the three features of lesser importance according to RF. Pretty obvious when you consider how industrial production and unemployment continue to deteriorate  for a few months after financial markets have begun their recovery. Possibly sarimax, but likely need interactions. Let's see if RNN plays nice with our small dataset.
