==> 3_LSTM_downturn_prediction.ipynb <==
import numpy as np
import matplotlib.pyplot as plt
import pandas as pd
import math
from keras.models import Sequential
from keras.layers import Dense,LSTM,BatchNormalization
from sklearn.preprocessing import MinMaxScaler
from sklearn.metrics import mean_squared_error

df = pd.read_csv('reg_ready.csv')
df['Date'] = pd.to_datetime(df['Date'])

REL_VARS = ['YIELD_CURVE','KO_diff_from_2_month_high',
       'KR_diff_from_2_month_high',
       'PPG_diff_from_2_month_high', 'SHW_diff_from_2_month_high',
       'unemployment_diff_from_6_month_low', 'INDPRO_diff_from_6_month_high']



df.dropna(subset=REL_VARS,inplace=True)
df.reset_index(inplace=True)
df.head()


#### Create windows of proper dimension (samples X timesteps X features)

X = []
y_downturn = []
y_monthly_return = []

for idx,dt in enumerate(df['Date']):
    obs_prior = df[df['Date']<=dt].shape[0]
    if obs_prior>=12:
        sample = []
        df_last_12 = df[df['Date']<=dt].iloc[-12:]
        sample.append(list(df_last_12['YIELD_CURVE']))
        sample.append(list(df_last_12['KO_diff_from_2_month_high']))
        sample.append(list(df_last_12['KR_diff_from_2_month_high']))
        sample.append(list(df_last_12['PPG_diff_from_2_month_high']))
        sample.append(list(df_last_12['SHW_diff_from_2_month_high']))
        sample.append(list(df_last_12['unemployment_diff_from_6_month_low']))
        sample.append(list(df_last_12['INDPRO_diff_from_6_month_high']))
        sample = np.array(sample).T
        
        X.append(sample)
        y_downturn.append(df.iloc[idx]['period of downturn'])
        y_monthly_return.append(df.iloc[idx]['monthly_return'])

arr1 = np.array(X)
arr1.shape

df_train = df.iloc[11:296]
print(df_train.shape)
df_test = df.iloc[297:]
print(df_test.shape)

#### Start with predicting monthly return on 1980 features

y = np.array(y_monthly_return)
X = arr1.copy()

trainX = X[:285,:,:]
trainY = y[:285]

testX = X[286:,:,:]
testY = y[286:]
print(trainX.shape)
print(testX.shape)  ##We can see that the dimensions of the numpy arrays are what we expected for merging with dfs.

# create and fit the LSTM network
batch_size = 1
features=7
timesteps=12



model = Sequential()
#model.add(LSTM(4, batch_input_shape=(batch_size, timesteps, features), return_sequences=True,activation='relu'))
#model.add(BatchNormalization())
model.add(LSTM(16, batch_input_shape=(batch_size, timesteps, features), activation='relu'))
#model.add(BatchNormalization())
model.add(Dense(1))
model.compile(loss='mean_squared_error', optimizer='adam')

model.summary()

results = model.fit(trainX, trainY, epochs=15, batch_size=batch_size, verbose=2, shuffle=False,validation_data=[testX,testY])


trainPredict = model.predict(trainX, batch_size=batch_size)

df_train['pred'] = trainPredict
df_train.to_csv('trainset_LSTM_1980.csv',index=False)

testPredict = model.predict(testX, batch_size=batch_size)

df_test['pred'] = testPredict
df_test.to_csv('testset_LSTM_1980.csv',index=False)

#### Let us try with downturn classification

y = np.array(y_downturn)
X = arr1.copy()

trainX = X[:285,:,:]
trainY = y[:285]

testX = X[286:,:,:]
testY = y[286:]
print(trainX.shape)
print(testX.shape)  ##We can see that the dimensions of the numpy arrays are what we expected for merging with dfs.

# create and fit the LSTM network
batch_size = 1
features=7
timesteps=12

model = Sequential()
#model.add(LSTM(4, batch_input_shape=(batch_size, timesteps, features), return_sequences=True,activation='relu'))
#model.add(BatchNormalization())
model.add(LSTM(16, batch_input_shape=(batch_size, timesteps, features), activation='relu'))
#model.add(BatchNormalization())
model.add(Dense(1,activation='sigmoid'))
model.compile(loss='binary_crossentropy', optimizer='adam')

results = model.fit(trainX, trainY, epochs=7, batch_size=batch_size, verbose=2, shuffle=False,validation_data=[testX,testY])

testPredict = model.predict(testX, batch_size=batch_size)

df_test['pred'] = testPredict
#df_test.to_csv('testset_LSTM_classification_1980.csv',index=False)

#### Let's try with fewer features - 1980, drop KO,KR, and monthly_return predicted

X = []
y_downturn = []
y_monthly_return = []

for idx,dt in enumerate(df['Date']):
    obs_prior = df[df['Date']<=dt].shape[0]
    if obs_prior>=12:
        sample = []
        df_last_12 = df[df['Date']<=dt].iloc[-12:]
        sample.append(list(df_last_12['YIELD_CURVE']))
        #sample.append(list(df_last_12['KO_diff_from_2_month_high']))
        #sample.append(list(df_last_12['KR_diff_from_2_month_high']))
        sample.append(list(df_last_12['PPG_diff_from_2_month_high']))
        sample.append(list(df_last_12['SHW_diff_from_2_month_high']))
        sample.append(list(df_last_12['unemployment_diff_from_6_month_low']))
        sample.append(list(df_last_12['INDPRO_diff_from_6_month_high']))
        sample = np.array(sample).T
        
        X.append(sample)
        y_downturn.append(df.iloc[idx]['period of downturn'])
        y_monthly_return.append(df.iloc[idx]['monthly_return'])
        
arr1 = np.array(X)
arr1.shape

df_train = df.iloc[11:296]
print(df_train.shape)
df_test = df.iloc[297:]
print(df_test.shape)

y = np.array(y_monthly_return)
X = arr1.copy()

trainX = X[:285,:,:]
trainY = y[:285]

testX = X[286:,:,:]
testY = y[286:]
print(trainX.shape)
print(testX.shape)  ##We can see that the dimensions of the numpy arrays are what we expected for merging with dfs.

# create and fit the LSTM network
batch_size = 1
features=5
timesteps=12



model = Sequential()

model.add(LSTM(16, batch_input_shape=(batch_size, timesteps, features), activation='relu'))
#model.add(BatchNormalization())
model.add(Dense(1))
model.compile(loss='mean_squared_error', optimizer='adam')

results = model.fit(trainX, trainY, epochs=15, batch_size=batch_size, verbose=2, shuffle=False,validation_data=[testX,testY])

#trainPredict = model.predict(trainX, batch_size=batch_size)
#df_train['pred'] = trainPredict
#df_train.to_csv('trainset_LSTM_1980_reducedfeats.csv',index=False)
#
#testPredict = model.predict(testX, batch_size=batch_size)
#df_test['pred'] = testPredict
#df_test.to_csv('testset_LSTM_1980_reducedfeats.csv',index=False)

#### So the 1980 full features is the best (or equivalent) of the RNN models tried. Let's comment out the extra to simplify file mgmt.

#### Not better (although more sensitive) than classification RF
